Made downloading input less verbose and fixed the tests.

+++ advent_of_code_template/advent_of_code_utils.py
@@ -1,6 +1,7 @@
 import pathlib
 
 from advent_of_code_template.definitions import TASKS_DIR
+from advent_of_code_template.definitions import CURRENT_YEAR
 import os
 from markdown import markdown
 import webbrowser
@@ -8,7 +9,6 @@ import importlib.util
 from typing import List
 import argparse
 import aocd
-from definitions import CURRENT_YEAR
 
 
 def parse_args(arguments: List[str]):
@@ -45,13 +45,17 @@ def download_input_data(task_number: int) -> List[str]:
         aocd.cookies.scrape_session_tokens()
     else:
         if not os.environ.get('AOC_SESSION'):
-            os.environ['AOC_SESSION'] = input('Please provide session ID in order to download the AOC data:')
             cookie_file = pathlib.Path('~/.config/aocd/token').expanduser()
-            print('Saving cookie to: ', cookie_file)
-            with open(cookie_file, 'w') as f:
-                f.write(os.environ['AOC_SESSION'])
-        else:
-            print('Downloading input-data using session ID: ', os.environ['AOC_SESSION'])
+            if os.path.exists(cookie_file):
+                print('Retrieving session ID from: ', cookie_file)
+                with open(cookie_file, 'r') as f:
+                    os.environ['AOC_SESSION'] = f.readline().strip()
+            else:
+                os.environ['AOC_SESSION'] = input('Please provide session ID in order to download the AOC data:')
+                print('Saving cookie to: ', cookie_file)
+                with open(cookie_file, 'w') as f:
+                    f.write(os.environ['AOC_SESSION'])
+        print('Downloading input-data for current task.')
     return aocd.get_data(day=task_number, year=CURRENT_YEAR).split('\n')
 
 
@@ -59,7 +63,6 @@ def get_input_data(task_number: int) -> List[str]:
     file_name = os.path.join(TASKS_DIR, 'task%02d/input.txt' % task_number)
     if not input_file_exists(task_number):
         task_input = download_input_data(task_number)
-        print('Successfully retrieved puzzle input. Saving input to: ', file_name)
         with open(file_name, 'w') as f:
             for line in task_input:
                 f.write(line)
